removed unused functions, renamed modules
