+++ minimdo/graph/workflow.py
@@ -1,5 +1,5 @@
 from graph.graphutils import dfs_tree, merge_edges, solver_children, end_components, Node, SOLVER, VAR, COMP, all_solvers, path
-from utils.utils import normalize_name
+from modeling.compute import normalize_name
 from collections import defaultdict
 from enum import Enum
