+++ minimdo/modeling/compute.py
@@ -79,4 +79,15 @@ class Par(Var):
         out = Var.__new__(self, *args, **kwargs)
         out.always_input = True
         return out
+
+
+def normalize_name(name, keep_underscore=False):
+    # f_{{{11}}} --> f11 
+    # \rho_{hi} --> rho_hi 
+    pattern = r'(\w+\_){(\w+)}'
+    sub1 = re.sub(pattern, r'\1\2', name)
+    if not keep_underscore:
+        sub1 = re.sub(r'\_', r'', sub1)
+    sub2 = re.sub(r'\\', r'', sub1)
+    return sub2
